Move SSM-related functions to extension-support
