+++ handel/src/common/rds-deployers-common.ts
@@ -17,7 +17,6 @@
 import { DeployContext, ServiceConfig, ServiceContext, UnDeployContext } from 'handel-extension-api';
 import { awsCalls } from 'handel-extension-support';
 import * as uuid from 'uuid';
-import * as ssmCalls from '../aws/ssm-calls';
 
 export function getDeployContext(serviceContext: ServiceContext<ServiceConfig>,
                                  rdsCfStack: any) { // TODO - Better type later
@@ -39,31 +38,6 @@ export function getDeployContext(serviceContext: ServiceContext<ServiceConfig>,
     });
 
     return deployContext;
-}
-
-// TODO - Once all logic using this is ported to TS, remove the "deployedStack" param
-export async function addDbCredentialToParameterStore(ownServiceContext: ServiceContext<ServiceConfig>,
-                                                      dbUsername: string,
-                                                      dbPassword: string,
-                                                      deployedStack: any) { // TODO - Better param later
-    // Add credential to EC2 Parameter Store
-    const usernameParamName = ownServiceContext.ssmParamName('db_username');
-    const passwordParamName = ownServiceContext.ssmParamName('db_password');
-    await Promise.all([
-        ssmCalls.storeParameter(usernameParamName, 'SecureString', dbUsername),
-        ssmCalls.storeParameter(passwordParamName, 'SecureString', dbPassword)
-    ]);
-    return deployedStack;
-}
-
-export async function deleteParametersFromParameterStore(ownServiceContext: ServiceContext<ServiceConfig>,
-                                                         unDeployContext: UnDeployContext) {
-    const paramsToDelete = [
-        ownServiceContext.ssmParamName('db_username'),
-        ownServiceContext.ssmParamName('db_password')
-    ];
-    await ssmCalls.deleteParameters(paramsToDelete);
-    return unDeployContext;
 }
 
 export function getNewDbUsername() {
