+++ handel/src/services/postgresql/index.ts
@@ -24,7 +24,7 @@ import {
     UnDeployContext,
     UnPreDeployContext
 } from 'handel-extension-api';
-import { awsCalls, bindPhase, deletePhases, handlebars, preDeployPhase, tagging } from 'handel-extension-support';
+import { awsCalls, bindPhase, deletePhases, deployPhase, handlebars, preDeployPhase, tagging } from 'handel-extension-support';
 import * as winston from 'winston';
 import * as rdsDeployersCommon from '../../common/rds-deployers-common';
 import { HandlebarsPostgreSQLTemplate, PostgreSQLConfig, PostgreSQLStorageType } from './config-types';
@@ -147,10 +147,7 @@ export async function deploy(ownServiceContext: ServiceContext<PostgreSQLConfig>
                                                                     30,
                                                                     stackTags);
         winston.debug(`${SERVICE_NAME} - Finished creating CloudFormation stack '${stackName}'`);
-        await rdsDeployersCommon.addDbCredentialToParameterStore(ownServiceContext,
-                                                                 dbUsername,
-                                                                 dbPassword,
-                                                                 deployedStack);
+        await deployPhase.addDbCredentialToParameterStore(ownServiceContext, dbUsername, dbPassword);
         winston.info(`${SERVICE_NAME} - Finished deploying database '${stackName}'`);
         return rdsDeployersCommon.getDeployContext(ownServiceContext, deployedStack);
     }
@@ -170,7 +167,8 @@ export function unBind(ownServiceContext: ServiceContext<PostgreSQLConfig>): Pro
 
 export async function unDeploy(ownServiceContext: ServiceContext<PostgreSQLConfig>): Promise<UnDeployContext> {
     const unDeployContext = await deletePhases.unDeployService(ownServiceContext, SERVICE_NAME);
-    return rdsDeployersCommon.deleteParametersFromParameterStore(ownServiceContext, unDeployContext);
+    await deletePhases.deleteParametersFromParameterStore(ownServiceContext);
+    return unDeployContext;
 }
 
 export const producedEventsSupportedServices = [];
